Revert "fixed bug related to log human locationi"
This reverts commit b87273cb
