+++ opendemic/human/location/controller.py
@@ -1,33 +1,45 @@
 from config.config import CONFIG, logger
 from flask import Blueprint, Response, render_template, request
 from opendemic.human import model
-from opendemic.human.model import Human
-from opendemic.human.location.types import LocationResourceFields
 from opendemic.human.location.util import Coordinate
 import json
+from enum import Enum
 
 blueprint = Blueprint('location', __name__)
 
 
+class LocationResourceFields(Enum):
+    FINGERPRINT = 'fingerprint'
+    LATITUDE = 'lat'
+    LONGITUDE = 'lng'
+    INCLUDE_LEGEND = 'include_legend'
+
+    @classmethod
+    def value_to_member_name(cls, value):
+        if cls.has_value(value):
+            return cls._value2member_map_[value].name
+
+    @classmethod
+    def has_value(cls, value):
+        return value in cls._value2member_map_
+
+
 @blueprint.route('/location', methods=['GET'])
 def location():
+    # Fetch or create human, log its location, and return nearby risky humans
     params = create_parameters()
     validation_error_response = validate_location_url_params(params=params)
     if len(validation_error_response) > 0:
         return_invalid_response(validation_error_response)
 
-    human: Human = model.get_human_from_fingerprint(fingerprint=params[LocationResourceFields.FINGERPRINT.value])
-    err: Exception = None
+    human = model.get_human_from_fingerprint(fingerprint=params[LocationResourceFields.FINGERPRINT.value])
     if human is None:
-        human, err = model.create_human(fingerprint=params[LocationResourceFields.FINGERPRINT.value])
-    if err is not None:
-        logger.error(err)
-    else:
-        human.log_location(
-            latitude=float(params[LocationResourceFields.LATITUDE.value]),
-            longitude=float(params[LocationResourceFields.LONGITUDE.value]),
-            send_alert=False
-        )
+        human = model.create_human(fingerprint=params[LocationResourceFields.FINGERPRINT.value])
+
+    human.log_location(
+        latitude=float(params[LocationResourceFields.LATITUDE.value]),
+        longitude=float(params[LocationResourceFields.LONGITUDE.value]),
+        send_alert=False)
 
     self_lat_lng = [params[LocationResourceFields.LONGITUDE.value], params[LocationResourceFields.LATITUDE.value]]
     self_geojson_feature = {
@@ -54,6 +66,7 @@ def location():
 
 
 def validate_location_url_params(params: dict) -> list:
+    # Sanity check for  location parameters
     validation_error_response = []
     for param_name in params:
         if params[param_name] is None:
@@ -88,6 +101,7 @@ def validate_location_url_params(params: dict) -> list:
 
 
 def create_parameters() -> dict:
+    # Create parameters from a request
     params = dict()
     params[LocationResourceFields.FINGERPRINT.value] = request.args.get(LocationResourceFields.FINGERPRINT.value)
     params[LocationResourceFields.LATITUDE.value] = request.args.get(LocationResourceFields.LATITUDE.value)
@@ -106,6 +120,7 @@ def create_parameters() -> dict:
 
 
 def return_invalid_response(message: list) -> Response:
+    # Return invalid response
     response = Response(
         response=json.dumps(message),
         status=403,
